Bug 2014218: Remove the kAOLMailMime type. r=mstange
Differential Revision: https://phabricator.services.mozilla.com/D281648

UltraBlame original commit: 6b665393fdf648736bd79d8ed29e68847539b8f0

+++ widget/nsPrimitiveHelpers.h
@@ -19,7 +19,7 @@ class nsPrimitiveHelpers {
   
   
   static void CreatePrimitiveForData(const nsACString& aFlavor,
-                                     const void* aDataBuff, uint32_t aDataLen,
+                                     const void* aDataBuff, size_t aDataLen,
                                      nsISupports** aPrimitive);
